First release of Power Settings
A pun on its ability to shutdown and restart your system, although more options were soon added. Making one add-on makes it easier to maintain.

It has no requirements.

FEATURES
- Shutdown your system
- Reboot your system
- Toggle fullscreen mode (pure javascript)
- Manually set the correct time

The last feature currently has an ability that warrent discussion. It can toggle whether the NTP daemon may use external servers to get the time. This is not a permanent setting (reboot and it's gone), it can only temporarily override this. It's fully optional.

I don't believe it hampers any of the other add-ons, since it's not like there will not be a time available. It also doesn't install anything new. It's just that by installing the add-on the user gains the ability to set the time manually. This has a few use cases.

A. Set the actual time
- Sometimes a user may want to override the time when the system is not providing the correct time. As long as NTP is allowed to update the time, users cannot set the time manually (except perhaps by fiddling with the timezone).

One example of this is when the system reboots when there is no internet connection available. I've been in this situation a lot. Exhibits, smart caravans that are not connected to the internet, etc.

B. Enhance privacy
- If a user disables the connection with NTP servers, privacy is enhanced. Currently, the NTP connection is one of the areas where there is little control over where the user's IP address gets set. While NTP is active, the user's system continuously emits requests for the current time, effectively sharing personal data (IP address) as well as behavioural data with unknown third parties. This add-on allows the WebThings Gateway to truly operate 'cloudlessly'.