Square theme 1.0.0
Adds log filtering feature, where uses can select which logs on the logs overview page they want to actually see.

This filtering list is toggled with a button in the top right corner of the screen. This bring up the question/suggestion: how about reserving that corner for add-on use?