Initial release of Followers
This addon allows one property to respond to a changing value of another. For example, as one value rises, another value could also be made to rise. Or it could be made to slowly decrease instead. 

It adds funtionality that cannot currently be easily created through the rules system.

If the rules system every gains this functionality, then this addon could be removed.

See also the "property to property" issue: https://github.com/mozilla-iot/gateway/issues/805