Initial proposal for Bluetooth Pairing addon
This extension addon does three things:
- allows users to scan for bluetooth devices
- click on the pairing button of a found device to pair it.
- once paired, the connect toggle appears

It was designed to make it easy to connect bluetooth speakers, but that has proven impossible without installing something like BlueAlsa or PulseAudio. It has been tested with Bluealsa and a small change to the bluetooth service file, and then it was indeed possible to connect to a bluetooth speaker. Despite connecting ok, I didn't hear any sound though.

Although untested, it should in theory still be possible to connect to mice and keyboards, but this is untested.

It is currently not possible to do things like passcodes.

As it could be useful to someone, and since the scanning functionality is already quite nice and the pairing works fine, I thought I'd release early.