Candle Hotspot - initial release
Initial release of the Candle hotspot addon. It is part of the Candle research for 2021. Work on this addon has been made possible by the generous support of the SIDN Foundation (https://www.sidnfonds.nl/excerpt)

This addon is only available for the Raspberry Pi.

With this addon the gateway will have:
- A very powerful and rare privacy feature. It brings some of the features of software like Dowse (https://www.dyne.org/software/dowse/) to the Webthings Gateway, in an incredible easy to use package.
- A way to run a gateway as a standalond "island". For example, you could disconnect the gateway from the network, and still connect to the gateway itself directly. For example with a dedicated tablet. This makes it very interesting for use in places like school, elderly homes and hotels, since the deployment could be guaranteed cloudless.
- A way to have a gateway act as a wifi range extender, this increasing the value proposition of the gateway.

It has been working flawlessly for months now. I also just tested it on a fresh gateway install, and it worked there too. Still, there a 90 second delay before the addon starts, just to be absolutely sure that the addon couldn't keep a user out of their system accidentally. In that period the user can use the interface to cancel the addon's launch (or just disable the addon in settings). It should be ample time to fix any issues if they somehow crop up. But as I said, it has been working very well here.

The next version will enable additional privacy protections, so that information about the connections made by laptops and mobile phones will be hidden.

I've made a lot of effort to make sure the addon is as safe to use as possible. Yes, this addon uses sudo internally in order to create an additional virtual network interface and launch hostapd. But it doesn't install anything onto the system, and it doesn't permanently change any settings of the host system. If a user removes the addon and reboots, it will be as if it had never been installed. Figuring out how to do this literally took weeks.